FRW-19 Refactoring compatibility traits for symfony 6
