+++ src/Silex/Provider/Routing/RedirectableUrlMatcherTrait.php
@@ -15,31 +15,7 @@ use Symfony\Component\HttpKernel\Kernel;
 
 // @phpstan-ignore-next-line
 if (Kernel::MAJOR_VERSION < 6) {
-    require 'RedirectableUrlMatcherTrait.symfony5.php';
-
-    return;
-}
-
-trait RedirectableUrlMatcherTrait
-{
-    /**
-     * @param string $path
-     * @param string $route
-     * @param string|null $scheme
-     *
-     * @return array
-     */
-    public function redirect(string $path, string $route, string $scheme = null): array
-    {
-        return $this->executeRedirect($path, $route, $scheme);
-    }
-
-    /**
-     * @param string $path
-     * @param string $route
-     * @param string|null $scheme
-     *
-     * @return array
-     */
-    abstract protected function executeRedirect(string $path, string $route, ?string $scheme = null): array;
+    require 'RedirectableUrlMatcherTrait/RedirectableUrlMatcherTrait.symfony5.php';
+} else {
+    require 'RedirectableUrlMatcherTrait/RedirectableUrlMatcherTrait.latest.php';
 }
